pyproject.toml
[project]
name = "lmaobot-hackey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
